Ensures that array values within an `ArraySerializable` are replaced
The fix introduced in #65 did a recursive merge of the data provided
when hydrating an `ArraySerializable` object. However, array _values_
should not be merged, as changes to them are generally not patches, but
rather replacements, as noted in #66.

This patch provides tests demonstrating the issues noted in #66, and
modifies the `ArraySerializable` hydration logic to use `array_merge()`
instead of `Zend\Stdlib\ArrayUtils::merge()`; all previous tests pass
now, as do the newly added tests.