Merge ArrayMapNamingStrategy into MapNamingStrategy
Per #9, the two perform almost identical functionality; the main
difference is _which argument is first in the constructor_, and the fact
that `MapNamingStrategy` allowed a second argument for the opposite
direction.

This patch merges the two into `MapNamingStrategy`. The class is now
final (as was `ArrayMapNamingStrategy`, and accepts either a
`$hydrationMap` or an `$extractionMap` or both. If both are missing, it
raises an exception. If one or the other is missing, the value becomes
that of the `array_flip` of the other value present.

The patch includes a migration document section detailing the changes,
and expands the `MapNamingStrategy` documentation to cover all use
cases.