Adds checks for arrays and array casting to MapNamingStrategy constructor
phpstan was complaining that the value passed to `flipMapping()` could
be null, and that the assignment to each of `extractionMap` and
`hydrationMap` could potentially be null but should not be. This patch
adds explicit `is_array()` checks before attempting to call
`flipMapping()` with a value, and casting to `array` during assignment
to properties.