Use named constructors
Removes the existing constructor, and creates three named constructors:

- `public static function createFromExtractionMap(array $extractionMap) : self`
- `public static function createFromHydrationMap(array $hydrationMap) : self`
- `public static function createFromAssymetricMap(array $extractionMap, array $hydrationMap) : self`

Doing so removes all the various constructor checks we were previously
using.

The patch also updates `flipMapping()` as follows:

- It requires string values _only_; integer are not allowed.
- It also checks the keys, as they can only be string values.

Finally, it updates the typehints on both the extraction and hydration
maps to `array<string, string`, as the keys are also important.