Initial standalone Hydrator plugin manager implementation
This patch provides an initial standalone plugin manager implementation,
for users who do not want to use zend-servicemanager, but still want to
use the `DelegatingHydrator`, or to compose the plugin manager in
classes that will retrieve hydrators.

To achieve this, it first adds the interface `Zend\Hydrator\HydratorPluginManagerInterface`,
which extends the PSR-11 `ContainerInterface`; this allows us to
typehint on hydrator plugin managers specifically, as well as ensure
they comply with PSR-11 (which zend-servicemanager v3 variants do not
currently). The `DelegatingHydratorFactory` now tests against that
interface instead of the `HydratorPluginManager` when attempting to
retrieve the hydrator plugin manager to use.

Next, the new `StandaloneHydratorPluginManager` is a non-extensible,
hard-coded container for retrieving the shipped hydrators only. If users
want something that can be configured, they should:

- Use the `HydratorPluginManager`
- Configure their application container to load hydrators
- Write their own implementation

The patch also adds a factory for the `StandaloneHydratorPluginManager`;
it simply instantiates and returns it.

Finally, `HydratorPluginManagerFactory` now raises an exception if
zend-servicemanager is not detected, and guides users to use the
`StandaloneHydratorPluginManager` or to install zend-servicemanager,
based on their configuration needs.