fix: Make CamelCaseToUnderscoreFilter behave as expected
Patch #14 inlined the `CamelCaseToUnderscore` and `UnderscoreToCamelCase`
filters in the package. However, in doing so, it added behavior that was
not expected, nor present in the original: splitting words on numeric
boundaries.

Furthermore, the functionality was unpredictable; some boundaries were
matched, while others were not, which led to an inability to predict
what would happen.

The problem stemmed from the addition of a regex in each of the unicode
and non-unicode matching pattern sets, as well as corresponding
replacement sets. These have been reverted to match those in
zend-filter, which restores behavior to how it worked in version 2.

Additional test cases were provided to the `UnderscoreToCamelCaseTest`
to demonstrate the reverse operations as well. One note: when converting
from underscore to camel case and word splits happen before numeric
values, there is no way to guarantee the opposite operation (camel case
to underscore) will resolve to the original value. If such input is
expected, it should likely use a custom filter.